Crate missing_why

Why missing

missing <-

When I was a Physics Ph.D. student in NYU in the late nineties, I took a course called UNIX tools in the CS department. It was a hands on course where the instructor did live REPL in the terminal and we watched him showing us all the tricks. I got hooked with UNIX since then. Got myself a dialup terminal in my tiny apartment in east village and dial in to the workstation on campus. The latency is so bad that I can’t only see the feedback after a few keystrokes. That was when I trained my vi muscle memory. (EMacs was out of the question.)

Later I got my own IBM 386 and installed Linux on it and started to program in Perl …

I am a big fan of Jon’s YouTube videos on Rust and I started to use Rust in non conventional ways.

I am going to follow this lecture series and “port” them to rustdoc and see how it goes.

What was missing then

Back then I only had a dumb terminal without GUI. Not enough memory or bandwidth to run IDEs like Emacs with sophisticated key binding and buffers. Even copy and paste was hard. So I have to type the command by hand, no tab completion.

Note: as a matter of fact, we don’t even need an editor to create text files, the text above was not inserted by an editor, although following is done via vi to save time:

cat <<EOF >> missing/missing-then.md
heredoc> Back then I only had a dumb terminal without GUI. Not enough memory
heredoc> or bandwidth to run IDEs like Emacs with sophisticated key binding and buffers.
heredoc> Even copy and paste was hard. So I have to type the command by hand, no tab completion. 
heredoc> 
heredoc> EOF

What is missing now

Our machines are orders of magnitude more powerful and we take orders of magnitude effort to accomplish the same tasks. We lost our muscle memory as well as real memory. We copy and paste but forgot what we have copied and where we copied from. We don’t stop and think about why and how.

More importantly, we let machine to automate the repetitive task and forget that repetition is the mother of learning.

Interactive computing is NOT WYSIWYG

Thirty years later, I realized that the excitement I experienced at the UNIX tools class was actually about interactive computing. Every keystroke I typed, every output the terminal printed, every mistake I made and fixed had huge impact on me.

Somehow along the way, this interactivity was replaced by the concept of WYSIWYG (What You See Is What You Get).

We moved from Typewriters to Word Processors. Editor as a profession becomes editor as a machine, just like Computer as a profession becomes computer as a machine.